extract OrcMapreduceRecordReader creation to util
